refactor: move template helpers from CppAnalyzer to SymbolIndexStore
Move _extract_template_base_name_from_usr (now static) and
find_template_specializations to SymbolIndexStore where they belong
(only access class_index, function_index, and index_lock).
CppAnalyzer retains backward-compatible facade methods.
