+++ mcp_server/symbol_index_store.py
@@ -6,6 +6,7 @@ and index maintenance operations.
 """
 
 import dataclasses
+import re
 from collections import defaultdict
 from typing import Any, Dict, List, Optional
 
@@ -281,3 +282,74 @@ class SymbolIndexStore:
             if file_path in self.file_index:
                 del self.file_index[file_path]
                 diagnostics.debug(f"Removed file {file_path} from file_index")
+
+    @staticmethod
+    def extract_template_base_name_from_usr(usr: str) -> Optional[str]:
+        """
+        Extract the base template name from a USR.
+
+        USR Format Examples:
+        - Generic Template:        c:@ST>1#T@Container
+        - Explicit Specialization: c:@S@Container>#I
+        - Partial Specialization:  c:@SP>1#T@Container>#*t0.0
+
+        Returns:
+            Base template name (e.g., "Container") or None if not a template-related USR
+        """
+        if not usr:
+            return None
+
+        match = re.search(r"c:@ST>[^@]*@(\w+)", usr)
+        if match:
+            return match.group(1)
+
+        match = re.search(r"c:@S@(\w+)", usr)
+        if match:
+            return match.group(1)
+
+        match = re.search(r"c:@SP>[^@]*@(\w+)", usr)
+        if match:
+            return match.group(1)
+
+        return None
+
+    def _add_class_template_symbols(self, base_name: str, results: List[SymbolInfo]) -> None:
+        """Add class template and specialization symbols to results."""
+        if base_name not in self.class_index:
+            return
+        for symbol in self.class_index[base_name]:
+            if symbol.kind in ("class_template", "partial_specialization"):
+                results.append(symbol)
+            elif symbol.kind in ("class", "struct"):
+                if symbol.usr and ">#" in symbol.usr:
+                    results.append(symbol)
+
+    def _add_function_template_symbols(self, base_name: str, results: List[SymbolInfo]) -> None:
+        """Add function template and specialization symbols to results."""
+        if base_name not in self.function_index:
+            return
+        for symbol in self.function_index[base_name]:
+            if symbol.kind == "function_template":
+                results.append(symbol)
+            elif symbol.kind in ("function", "method"):
+                if symbol.is_template_specialization or (
+                    symbol.usr and ("<#" in symbol.usr or ">#" in symbol.usr)
+                ):
+                    results.append(symbol)
+
+    def find_template_specializations(self, base_name: str) -> List[SymbolInfo]:
+        """
+        Find all specializations of a template by base name.
+
+        Searches for:
+        1. Generic template definition (kind=class_template, function_template)
+        2. Explicit full specializations (kind=class, function with template args in USR)
+        3. Partial specializations (kind=partial_specialization)
+        """
+        results: List[SymbolInfo] = []
+
+        with self.analyzer.index_lock:
+            self._add_class_template_symbols(base_name, results)
+            self._add_function_template_symbols(base_name, results)
+
+        return results
